move around the tests to organize by cipher
