+++ tests/hazmat/primitives/test_camellia.py
@@ -12,19 +12,38 @@
 # limitations under the License.
 
 """
-Test using the OpenSSL Test Vectors
+Tests using the CRYPTREC (Camellia) Test Vectors
 """
 
 from __future__ import absolute_import, division, print_function
 
 import binascii
-
 import os
 
 from cryptography.hazmat.primitives.block import ciphers, modes
 
 from .utils import generate_encrypt_test
-from ...utils import load_openssl_vectors_from_file
+from ...utils import (
+    load_cryptrec_vectors_from_file, load_openssl_vectors_from_file
+)
+
+
+class TestCamelliaECB(object):
+    test_NTT = generate_encrypt_test(
+        load_cryptrec_vectors_from_file,
+        os.path.join("ciphers", "Camellia"),
+        [
+            "camellia-128-ecb.txt",
+            "camellia-192-ecb.txt",
+            "camellia-256-ecb.txt"
+        ],
+        lambda key: ciphers.Camellia(binascii.unhexlify((key))),
+        lambda key: modes.ECB(),
+        only_if=lambda backend: backend.ciphers.supported(
+            ciphers.Camellia("\x00" * 16), modes.ECB()
+        ),
+        skip_message="Does not support Camellia ECB",
+    )
 
 
 class TestCamelliaCBC(object):
@@ -66,18 +85,4 @@ class TestCamelliaCFB(object):
             ciphers.Camellia("\x00" * 16), modes.CFB("\x00" * 16)
         ),
         skip_message="Does not support Camellia CFB",
-    )
-
-
-class TestAESCTR(object):
-    test_OpenSSL = generate_encrypt_test(
-        load_openssl_vectors_from_file,
-        os.path.join("ciphers", "AES", "CTR"),
-        ["aes-128-ctr.txt", "aes-192-ctr.txt", "aes-256-ctr.txt"],
-        lambda key, iv: ciphers.AES(binascii.unhexlify(key)),
-        lambda key, iv: modes.CTR(binascii.unhexlify(iv)),
-        only_if=lambda backend: backend.ciphers.supported(
-            ciphers.AES("\x00" * 16), modes.CTR("\x00" * 16)
-        ),
-        skip_message="Does not support AES CTR",
     )
